ThreatTableComponent
Revamped the threat table component in C++. Added a lot of helper functions and formulas for dealing/taking/healing damage.
Tick function for VE reduction.

New interface functions for ThreatTableComponent and XIAggroComponent.

New Character Interface function to set a main target. Will need to override it in the base enemy to include applying the tag for combat stance.