JobStatsExecCalc + misc
Base stats are now applied to JobStatsExecCalc which is called when a Hero changes their job. (called from that ability).

Set a character's level according to the specified GE in the job ability.
TODO: Add the level check for the job ability to only add ones the character is at level for.