EquipmentWidget Progress
Nearly finished the equipment widget. You are able to navigate the widget go to an inventory widget and equip items and see item details on hover.

The next objective is to be able to sort/filter the items visible based on equipslot. That can be tackled at a later date.